+++ packages/framework/view-interfaces/src/index.ts
@@ -3,5 +3,5 @@
  * Licensed under the MIT License.
  */
 
-export * from "./componentRender";
-export * from "./interfaces";
+export * from "./htmlView";
+export * from "./reactView";
